some moving around (1)
