+++ Lib/Android/FreePIE Android IMU/src/com/freepie/android/imu/dsp/Biquad.java
@@ -1,4 +1,4 @@
-package com.freepie.android.imu;
+package com.freepie.android.imu.dsp;
 
 public class Biquad implements Filter {
 	public final static int LOWPASS = 0;
