+++ Lib/Android/FreePIE Android IMU/src/com/freepie/android/imu/dsp/Filter.java
@@ -1,4 +1,4 @@
-package com.freepie.android.imu;
+package com.freepie.android.imu.dsp;
 
 public interface Filter {
 	double filter(double x);
